rename legacy class package
