+++ pysiral/core/__init__.py
@@ -7,7 +7,7 @@ __author__ = "Stefan Hendricks"
 
 __all__ = [
     "class_template", "clocks", "config", "datahandler", "errorhandler",
-    "flags", "helper", "iotools", "legacy_packages", "output",
+    "flags", "helper", "iotools", "legacy_classes.py", "output",
     "DefaultLoggingClass", "functions"
 ]
