+++ pysiral/core/legacy_classes.py
