+++ pysiral/core/output.py
@@ -10,7 +10,7 @@ from pathlib import Path
 import cftime
 import numpy as np
 import parse
-from pysiral.core.legacy_packages import AttrDict
+from pysiral.core.legacy_classes import AttrDict
 from dateutil import parser as dtparser
 from loguru import logger
 from netCDF4 import Dataset, date2num
