+++ pysiral/grid.py
@@ -11,7 +11,7 @@ from typing import Any, Tuple, Union
 import numpy as np
 import numpy.typing as npt
 import scipy.ndimage as ndimage
-from pysiral.core.legacy_packages import AttrDict
+from pysiral.core.legacy_classes import AttrDict
 from pyproj import Proj
 from pyresample import geometry
