+++ pysiral/l1preproc/procitems.py
@@ -52,7 +52,7 @@ are currently implemented:
 
 from typing import Any, Dict, List, TypeVar, Union
 
-from pysiral.core.legacy_packages import AttrDict
+from pysiral.core.legacy_classes import AttrDict
 from schema import And, Schema
 
 from pysiral import get_cls
