+++ tests/test_definition_files.py
@@ -9,7 +9,7 @@ import datetime
 import unittest
 from pathlib import Path
 
-from pysiral.core.legacy_packages import AttrDict
+from pysiral.core.legacy_classes import AttrDict
 from loguru import logger
 
 from pysiral import psrlcfg
