+++ tests/test_l2_ouput_def_files.py
@@ -10,7 +10,7 @@ Level2Processor conventions
 
 import unittest
 
-from pysiral.core.legacy_packages import AttrDict
+from pysiral.core.legacy_classes import AttrDict
 from loguru import logger
 
 from pysiral import psrlcfg
